+++ changelog.md
@@ -5,9 +5,11 @@ I just really wanted to make a changelog. I might make this into a library or so
 All changes to this project from 2024-04-06 onwards are documented here.<hr>
 ## (Unreleased) [1.1.0] 2024-04-07<hr>
 ### Added
-
+* GUI for selecting atoms.
+* Boolean for atom rotation
 ### Changed
 * Modularized Javascript code into app components, manager components, and mesh/body components.
+* Separated `AtomManager` class into `BohrAtomManager` and `QuantumAtomManager` classes.
 ### Removed
 ## [1.0.0] 2024-04-06<hr>
 ### Added